+++ lib/app/wenyou_app.dart
@@ -7,6 +7,7 @@ import 'package:wenyousite_mobile/app/app_theme.dart';
 import 'package:wenyousite_mobile/core/widgets/wenyou_instant_keyboard_insets.dart';
 import 'package:wenyousite_mobile/features/app_shell/application/startup_controller.dart';
 import 'package:wenyousite_mobile/features/app_shell/presentation/startup_gate.dart';
+import 'package:wenyousite_mobile/features/media/application/avatar_image_ports.dart';
 import 'package:wenyousite_mobile/features/media/application/media_upload_task_controller.dart';
 import 'package:wenyousite_mobile/features/media/data/editor_image_picker.dart';
 import 'package:wenyousite_mobile/features/media/data/media_upload_repository.dart';
@@ -21,6 +22,9 @@ class WenyouApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return ProviderScope(
       overrides: [
+        avatarImagePickerPortProvider.overrideWith(
+          (ref) => ref.watch(avatarImagePickerProvider),
+        ),
         editorImagePickerPortProvider.overrideWith(
           (ref) => ref.watch(editorImagePickerProvider),
         ),